Module mio_extras::channel

Thread safe communication channel implementing Evented

Structs

Receiver
The receiving half of a channel.

Sender
The sending half of a channel.

SyncSender
The sending half of a synchronous channel.

Enums

SendError
An error returned from the Sender::send or SyncSender::send function.

TrySendError
An error returned from the SyncSender::try_send function.

Functions

channel
Creates a new asynchronous channel, where the Receiver can be registered with Poll.

sync_channel
Creates a new synchronous, bounded channel where the Receiver can be registered with Poll.